Tau event generator (#6)
This commit adds tau generation to the event generator in an efficient way. Vertices in a large volume are being generated but only interactions in a smaller fiducial volume are being saved. For more details see discussions on pull request

* remove 'new' event file format that will not be adapted

* add functionality to hdf5 writer functions to write a second dataset table

* modify event generator to add some dummy tau second bang interactions,
add hdf5 testing scripts

* Physics of the tau decay. A beginning.

* Changed an argument.

* fix typo

* insert second vertices into main table

* instead of adding a new table with additional vertices, insert additional vertices into the first table as new rows.
the hdf5 writer is adjusted to not split a file in the middle of an event. Also the 'number of events' attribute is set correctly to count the number of event and NOT the number of vertices in a file

* Output files with the same number of events each.

* Refactorisation. Tau energy is now correctly calculated using the inelasticity.

* Solved problem that happened while merging.

* save tau decaus only if decay takes place in fiducial volume

* add first step to generate tau decays from nu_tau interactions outside of fiducial volume

* add .settings folder to gitignore

* implement computationally efficient tau second band generation (details will be discussed on pull request)

* add debug output for tau decays