Fix typo in raytraycing and also fix print statements to be py3 compatible (#206)
This commit modifies three files where there is an typo calling raytraycing, which is an old version of the ray tracer which had typos. The module name was fixed, but the change wasn't propagated to the examples apparently.

Also, three print statements only worked in py2 in one of the examples, so those are now py3 compatible.