Add check_modules script
This script will check if all run methods defined in a class in the
module directory have been registered. It takes advantage of the fact
that a time attribute is added to the run function by the decorator.

Since each module must be imported as part of this check, it also as a
byproduct finds broken modules (though they may only be broken for the
user running it, if e.g. an optional dependency is required).

Probably this is not the right place for the script but I didn't want to
figure out relative paths :)